requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0

# Kraken
krakenex>=2.2.0
pyjwt>=2.8.0

# Blockchain (ERC-8004)
web3>=6.15.0
eth-typing>=4.0.0

# AI
openai>=1.12.0  # Compatible with Nous API

# Logging & Utils
python-json-logger>=2.0.7
schedule>=1.2.0
